fix(ci): cast c_long return to i64

+++ crates/atlas-runtime/src/ffi/caller.rs
@@ -129,7 +129,7 @@ impl ExternFunction {
             }
             "()->CLong" => {
                 let f: extern "C" fn() -> c_long = std::mem::transmute(self.fn_ptr);
-                Ok(CType::Long(f()))
+                Ok(CType::Long(i64::from(f())))
             }
             "()->CDouble" => {
                 let f: extern "C" fn() -> c_double = std::mem::transmute(self.fn_ptr);
